speech/requirements.txt
faster-whisper>=0.10.0

# Ollama integration (SSH tunneling)
sshtunnel>=0.4.0
paramiko>=2.12.0,<3.4
requests>=2.28.0

# BigQuery integration
google-cloud-bigquery>=3.21.0

# Optional: for GPU support
# torch>=2.0.0
# Optional: for audio resampling (if not using system ffmpeg)
# librosa>=0.10.0

